coreinit: Refactor some memory heap stuff.

+++ src/libdecaf/src/modules/coreinit/coreinit_frameheap.h
@@ -21,7 +21,7 @@ MEMCreateFrmHeap(FrameHeap *heap,
 FrameHeap *
 MEMCreateFrmHeapEx(FrameHeap *heap,
                    uint32_t size,
-                   uint16_t flags);
+                   uint32_t flags);
 
 void *
 MEMDestroyFrmHeap(FrameHeap *heap);
